feat: race future and ffmpeg, pause if ffmpeg doesn't finish
In the future we will race user input, and the ffmpeg process. If user
input occurs, we will wait until a certain time period passes before
we resume ffmpeg.

This commit introduces the struct for pausing / unpausing processes,
`PausableProcess`, a struct which pauses on Drop, `PauseOnDrop`, and
a little demo of it working in main.rs by racing a sleep and ffmpeg.