feat: initial configuration for video paths.
This initial configuration only supports adding in where your videos
are, and where the encoded videos should go. Right now, it doesn't
have any way of e.g. detecting `~`, because windows won't just turn
that into your user directory nicely for you. I use directories to
grab the Videos folder right now as a default, and throw it in an
encoded folder, but I'm not sure on what the exact behavior here should
be yet.